Add: Customize groups for Ement faces
Each face now belongs to one of the following new groups:

- ement-room-faces (sub-group of ement-room)
- ement-room-list-faces (sub-group of ement-room-list)
- ement-tabulated-room-list-faces (sub-group of ement-tabulated-room-list)

Additionally, each of those new groups is also a sub-group of another new
group 'ement-faces' (which is parent to only the three 'faces' sub-groups).

You can therefore navigate to each list of faces from either the customize
group the faces used to belong to, or via the new 'ement-faces' container.

Also:

- Delete duplicate defgroup ement-room-list.
- Tweak phrasing for defgroup ement-tabulated-room-list for
  consistency with ement-room-list.